Hago parte de la memoria y arreglo el ES

+++ include/ES.h
@@ -18,7 +18,7 @@
  * @param generator generador de números aleatorios
  * @return double temperatura tras enfriamiento
  */
-double Enfriamiento (double T, double T_inicial, double T_final, int M, std::mt19937 &generator);
+double Enfriamiento (double T, double T_inicial, double T_final, double M, std::mt19937 &generator);
 
 /**
  * @brief Calcula la temperatura inicial
